+++ accepton/api/querying.py
@@ -56,6 +56,28 @@ class Querying(Utils):
                                             {},
                                             PromoCode)
 
+    def promo_codes(self, order=None, order_by=None, page=None, per_page=None,
+                    promo_type=None):
+        """Retrieves a page of promo codes from the API
+
+        :param order: The order to sort by (asc or desc).
+        :type order: str.
+        :param order_by: The field to order by (e.g. created_at).
+        :type order_by: str.
+        :param page: The page number to retrieve.
+        :type page: int.
+        :param per_page: The size of the page to retrieve (max: 100).
+        :type per_page: int.
+        :param promo_type: The type of promo code to filter by.
+        :type promo_type: str.
+
+        :returns: List<PromoCode> -- The list of retrieved PromoCodes.
+        :raises: accepton.Error
+        """
+        return self.perform_get_with_objects("/v1/promo_codes",
+                                             self.as_params(locals()),
+                                             PromoCode)
+
     def token(self, id):
         """Retrieves a transaction token from the API