+++ CHANGELOG.rst
@@ -10,7 +10,10 @@ this scheme are considered to be bugs.
 `Unreleased`_
 -------------
 
-- Nothing yet.
+Added
+~~~~~
+
+- Added support for querying lists of promo codes.
 
 0.1.0 - 2015-07-20
 ------------------